Do not fail if an ObjStm ends with a number
Without this fix, pdftk fails for example on particular PDF version
1.5 documents generated by newer pdflatex versions.

The actual bug is in the method PRTokeniser#nextValidToken. This
method has the feature that it treats an indirect object reference
(such as 24 0 R) as a single token. Therefore when it sees a number,
it has to look ahead to see if the number is actually the start of an
indirect object reference.  If, however, the object stream ends during
this lookahead then it would wrongly fail. So it will go wrong
whenever the last object in an object stream is a number.

Author: Robin Houston <robin.houston@gmail.com>
Bug-Debian: http://bugs.debian.org/687669

[W. Trevor King]: This is the Debian commit message for Robin's
963ed828156fc7cd29366289638bb1b820b12210